Reduced threshold for warnings in codebeat
Codebeat (rightly so) raises many warnings about the number of “instance variables” in a struct being too high.
As this library is performing unmarshalling of raw json strings into golang strings from the GoCD api, the amount of flexibility in changing those structs is limited.
n option of writing custom unmarshallers to break the structs down, is an option, but I think this would be undue complexity added to the library.